Reports sorted out.

+++ application/models/Database.php
@@ -3,7 +3,7 @@ class Database extends CI_Model
 {
     public function getNode($id)
     {
-        $result = $this->db->where('id_node', $id)->get('node')->result_array();
+        $result = $this->db->where('id_node', $id)->where('reports <', 3)->get('node')->result_array();
         if (count($result))
         {
             $this->db->where('id_node', $id)->update('node', array('clicks'=>$result[0]['clicks'] + 1));
@@ -35,7 +35,7 @@ class Database extends CI_Model
 
     public function getOptions($id)
     {
-        $result = $this->db->where('source_node', $id)->get('node')->result_array();
+        $result = $this->db->where('source_node', $id)->where('reports <', 3)->get('node')->result_array();
         shuffle($result);
         $result = array_slice($result, 0, min(count($result), 3));
         foreach($result as $res)
@@ -53,30 +53,32 @@ class Database extends CI_Model
 
     public function doesIdExist($id)
     {
-        $result = $this->db->where('source_node',$id)->get('node')->result_array();
+        $result = $this->db->where('id_node',$id)->get('node')->result_array();
         return count($result) > 0;
     }
 
     public function report($id)
     {
-        $result = $this->db->where('id_node', $id)->get('node')->result_array();
+        $result = $this->db->where('id_node', $id)->where('reports <', 3)->get('node')->result_array();
         if (count($result) && $result[0]['reports'] >= 0)
         {
-            $this->db->where('id_node',$result[0]['id_node'])->update('node',array('reports'=>$result[0]['reports'] + 1));
+            $this->db->where('id_node',$result[0]['id_node'])->update('node', array('reports'=>$result[0]['reports'] + 1));
         }
     }
 
     public function removeFreeNodes()
     {
         $count = 0;
+
         foreach($this->db->get('node')->result_array() as $node)
         {
-            if(!doesIdExist($node['source_node']))
+            if($node['source_node'] > 0 && !$this->doesIdExist($node['source_node']))
             {
                 $this->db->where('id_node', $node['id_node'])->delete('node');
                 $count++;
             }
         }
+
         return $count;
     }
 }
